+++ mobile/package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "@expo/ngrok": "^4.1.0",
     "@expo/webpack-config": "^0.17.2",
-    "@react-native-community/datetimepicker": "6.2.0",
     "@react-navigation/native": "^6.0.11",
     "@react-navigation/native-stack": "^6.8.0",
     "@react-navigation/stack": "^6.2.2",
