+++ mobile/screens/AnimalInformationScreen.tsx
@@ -1,34 +1,7 @@
 import { StatusBar } from "expo-status-bar";
 import React, { useState } from "react";
-import { useWindowDimensions, StyleSheet, Text, View } from "react-native";
-import { TextInput, TouchableOpacity } from "react-native-gesture-handler";
-
-function validateDate(dateStr: string) {
-  const regex = /^\d{1,2}-\d{1,2}-\d{4}$/;
-
-  if (!regex.test(dateStr)) {
-    return false;
-  }
-
-  // Parse the date parts to integers
-  var parts = dateStr.split("-");
-  var day = parseInt(parts[1]);
-  var month = parseInt(parts[0]);
-  var year = parseInt(parts[2]);
-
-  // Check the ranges of month and year
-  if (year < 1000 || year > 3000 || month == 0 || month > 12)
-    return false;
-
-  var monthLength = [ 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 ];
-
-  // Adjust for leap years
-  if (year % 400 == 0 || (year % 100 != 0 && year % 4 == 0))
-    monthLength[1] = 29;
-
-  // Check the range of the day
-  return day > 0 && day <= monthLength[month - 1];
-}
+import { StyleSheet, Text, View, TextInput, TouchableOpacity } from "react-native";
+import { validateDate } from "../utils/string";
 
 function updateDatabase() {
 
@@ -41,8 +14,6 @@ export default function AnimalInformationScreen(props: any) {
   const [animalBirth, setAnimalBirth] = useState("");
   const [animalAdaptation, setAnimalAdaptation] = useState("");
   const [error, setError] = useState("");
-
-  const windowHeight = useWindowDimensions().height;
 
   return (
     <View
@@ -183,6 +154,7 @@ const styles = StyleSheet.create({
     alignSelf: "flex-start",
   },
   failedContainer: {
+    marginTop: 12,
     alignItems: "center",
     marginBottom: 12,
     padding: 8,
