+++ package.json
@@ -22,7 +22,6 @@
     "start:mobile-tunnel": "cd mobile && expo start --tunnel"
   },
   "dependencies": {
-    "concurrently": "^7.3.0",
-    "react-native-keyboard-aware-scroll-view": "^0.9.5"
+    "concurrently": "^7.3.0"
   }
 }
